requirements.txt
Pillow==10.4.0
ttkbootstrap==1.14.1
